Added feature_extraction.py into predict.py
predict.py now does feature extraction on the EEG raw data before predicting using the model. It now returns a tuple of the probability distribution of the predicted emotion as well as the predicted emotion.

The main function gives an example on how to use the predict function on SEED IV raw EEG data.